refactor: copy기능 분리
