+++ src/common/modal/Modal.module.css
@@ -4,7 +4,6 @@
 모달은 화면의 수평, 수직 가운데로 배치해 주세요. */
 
 .container {
-  border: 5px solid blue;
   position: fixed;
   top: 0;
   left: 0;
